Implement new Firebase authentication logic
