+++ src/App.tsx
@@ -2,7 +2,7 @@ import { useCallback, useEffect } from "react"
 import { useLocation } from "react-router-dom"
 import AddToWatchlistModal from "./components/AddToWatchlistModal/AddToWatchlistModal"
 import AppRoutes from "./components/AppRoutes"
-import { useAuth } from "./context/AuthContext"
+import { useGetUserQuery } from "./services/firebase/auth/queries"
 
 function App() {
   // Handle the link click to scroll to top
@@ -16,7 +16,7 @@ function App() {
     handleLinkClick()
   }, [pathname, handleLinkClick])
 
-  const { currentUser } = useAuth()
+  const { data: currentUser } = useGetUserQuery()
 
   return (
     <>
