+++ src/index.tsx
@@ -4,7 +4,6 @@ import React from "react"
 import ReactDOM from "react-dom/client"
 import { BrowserRouter } from "react-router-dom"
 import App from "./App"
-import { AuthProvider } from "./context/AuthContext"
 import { HistoryContextProvider } from "./context/HistoryContext"
 import { ModalContextProvider } from "./context/ModalContext"
 import { SidebarContextProvider } from "./context/SidebarContext"
@@ -17,18 +16,16 @@ ReactDOM.createRoot(document.getElementById("root") as HTMLDivElement).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
       <UserProvider>
-        <AuthProvider>
-          <ModalContextProvider>
-            <HistoryContextProvider>
-              <SidebarContextProvider>
-                <BrowserRouter>
-                  <App />
-                  <ReactQueryDevtools initialIsOpen={true} />
-                </BrowserRouter>
-              </SidebarContextProvider>
-            </HistoryContextProvider>
-          </ModalContextProvider>
-        </AuthProvider>
+        <ModalContextProvider>
+          <HistoryContextProvider>
+            <SidebarContextProvider>
+              <BrowserRouter>
+                <App />
+                <ReactQueryDevtools initialIsOpen={true} />
+              </BrowserRouter>
+            </SidebarContextProvider>
+          </HistoryContextProvider>
+        </ModalContextProvider>
       </UserProvider>
     </QueryClientProvider>
   </React.StrictMode>
