--- src/services/firebase/auth/mutationsR1.ts
@@ -1,50 +1,0 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { useNavigate } from "react-router-dom"
-import { useNavigateBack } from "../../../hooks/useNavigateBack"
-import {
-  authWithEmailAndPassword,
-  signInWithProvider,
-  signOut,
-} from "./endPoints"
-
-export const useSignInWithProviderMutation = () => {
-  const navigateBack = useNavigateBack()
-  const queryClient = useQueryClient()
-  return useMutation({
-    mutationFn: signInWithProvider,
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: ["currentUser"],
-      })
-      navigateBack()
-    },
-  })
-}
-
-export const useAuthWithEmailAndPasswordMutation = () => {
-  const navigateBack = useNavigateBack()
-  const queryClient = useQueryClient()
-  return useMutation({
-    mutationFn: authWithEmailAndPassword,
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: ["currentUser"],
-      })
-      navigateBack()
-    },
-  })
-}
-
-export const useLogoutMutation = () => {
-  const navigate = useNavigate()
-  const queryClient = useQueryClient()
-  return useMutation({
-    mutationFn: signOut,
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: ["currentUser"],
-      })
-      navigate("/")
-    },
-  })
-}
